requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
python-dotenv==1.0.0
pyTelegramBotAPI==4.10.0
Werkzeug==2.2.3
requests==2.26.0
